pyproject.toml
[project]
name = "project-8-equipo1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "imbalanced-learn>=0.14.1",
    "matplotlib>=3.10.8",
    "mlflow>=2.12.0,<3.0.0",
    "numpy>=2.4.4",
    "optuna>=4.8.0",
    "pandas>=2.0.0,<3.0.0",
    "protobuf<4",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "setuptools>=82.0.1",
    "stats>=0.1.2a0",
    "statsmodels>=0.14.6",
    "streamlit>=1.56.0",
    "supabase>=2.28.3",
    "xgboost>=3.2.0",
]
